+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 ## [Unreleased]
+### Changed
+- Client version updated on [5.3.15](https://github.com/reportportal/client-java/releases/tag/5.3.15), by @HardNorth
 
 ## [5.3.6]
 ### Fixed
 - Hook suites are now correctly inherit "SKIPPED" status from child items, by @HardNorth
+### Changed
+- Client version updated on [5.3.14](https://github.com/reportportal/client-java/releases/tag/5.3.14), by @HardNorth
 
 ## [5.3.5]
 ### Removed